requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-gis==1.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
